feat: use new compute unit instructions (#3)
* feat: use new compute unit program insturction

* feat: both of devent and testnet use compute unit tx

+++ config.go
@@ -29,7 +29,7 @@ type PingConfig struct {
 	StatusCheckInterval     int64
 	MinPerPingTime          int64
 	RequestUnits            uint32
-	ComputeUnitPrice        uint32
+	ComputeUnitPrice        uint64
 }
 
 type SlackReport struct {
